# mate-platform-backend/contracts/openapi/generated/bundled.yaml
openapi: 3.1.0
info:
  title: Mate Platform API
  version: 1.0.0
  description: Canonical contract-first API for all Mate Platform domains.
  contact:
    name: Mate Platform API Governance
    email: api-governance@metaplatform.local
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-mate-owner: architecture-api-governance
servers:
  - url: /
    description: Gateway-relative production endpoint
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/etl/tasks/{id}:
    get:
      tags:
        - data
      summary: 详情 ETL任务
      operationId: dataGetEtlTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 详情 ETL任务
      x-mate-owner: data-platform
      x-mate-permission: data.get
      x-mate-requirements:
        - FR-DATA-DATAGETETLTASKSID
      x-mate-implementation-status: planned
    put:
      tags:
        - data
      summary: 更新 ETL任务
      operationId: dataPutEtlTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 更新 ETL任务
      x-mate-owner: data-platform
      x-mate-permission: data.put
      x-mate-requirements:
        - FR-DATA-DATAPUTETLTASKSID
      x-mate-implementation-status: planned
    delete:
      tags:
        - data
      summary: 删除 ETL任务
      operationId: dataDeleteEtlTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 删除 ETL任务
      x-mate-owner: data-platform
      x-mate-permission: data.delete
      x-mate-requirements:
        - FR-DATA-DATADELETEETLTASKSID
      x-mate-implementation-status: planned
...
  /api/v1/scheduler/tasks/{id}:
    get:
      tags:
        - data
      summary: 详情 调度任务
      operationId: dataGetSchedulerTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 详情 调度任务
      x-mate-owner: data-platform
      x-mate-permission: data.get
      x-mate-requirements:
        - FR-DATA-DATAGETSCHEDULERTASKSID
      x-mate-implementation-status: planned
    put:
      tags:
        - data
      summary: 更新 调度任务
      operationId: dataPutSchedulerTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 更新 调度任务
      x-mate-owner: data-platform
      x-mate-permission: data.put
      x-mate-requirements:
        - FR-DATA-DATAPUTSCHEDULERTASKSID
      x-mate-implementation-status: planned
    delete:
      tags:
        - data
      summary: 删除 调度任务
      operationId: dataDeleteSchedulerTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      description: 删除 调度任务
      x-mate-owner: data-platform
      x-mate-permission: data.delete
      x-mate-requirements:
        - FR-DATA-DATADELETESCHEDULERTASKSID
      x-mate-implementation-status: planned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          enum:
            - 0
        message:
          type: string
        data: {}
        traceId:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
        traceId:
          type: string
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...